cpio: use uint64 for Mknod on FreeBSD
see also https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=236131

+++ pkg/cpio/fs_unix.go
@@ -159,12 +159,12 @@ func CreateFileInRoot(f Record, rootDir string, forcePriv bool) error {
 		}
 
 	case os.ModeDevice:
-		if err := syscall.Mknod(f.Name, perm(f)|syscall.S_IFBLK, dev(f)); err != nil && forcePriv {
+		if err := syscall.Mknod(f.Name, perm(f)|syscall.S_IFBLK, uint64(dev(f))); err != nil && forcePriv {
 			return err
 		}
 
 	case os.ModeCharDevice:
-		if err := syscall.Mknod(f.Name, perm(f)|syscall.S_IFCHR, dev(f)); err != nil && forcePriv {
+		if err := syscall.Mknod(f.Name, perm(f)|syscall.S_IFCHR, uint64(dev(f))); err != nil && forcePriv {
 			return err
 		}
